+++ example/lib/main.dart
@@ -1,12 +1,14 @@
 import 'dart:async';
 
-import 'package:flare_flutter/flare_actor.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+
 import 'package:pin_code_fields/pin_code_fields.dart';
+
+import './constants/constants.dart';
 
 void main() => runApp(MyApp());
 
+// ignore: use_key_in_widget_constructors
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
@@ -15,17 +17,23 @@ class MyApp extends StatelessWidget {
       title: 'Flutter Demo',
       theme: ThemeData(
         primarySwatch: Colors.blue,
+        backgroundColor: Colors.white,
+        scaffoldBackgroundColor: Colors.white,
       ),
-      home: PinCodeVerificationScreen(
-          "+8801376221100"), // a random number, please don't call xD
+      home: const PinCodeVerificationScreen(
+          phoneNumber:
+              "+8801376221100"), // a random number, please don't call xD
     );
   }
 }
 
 class PinCodeVerificationScreen extends StatefulWidget {
-  final String phoneNumber;
+  final String? phoneNumber;
 
-  PinCodeVerificationScreen(this.phoneNumber);
+  const PinCodeVerificationScreen({
+    Key? key,
+    this.phoneNumber,
+  }) : super(key: key);
 
   @override
   _PinCodeVerificationScreenState createState() =>
@@ -33,60 +41,61 @@ class PinCodeVerificationScreen extends StatefulWidget {
 }
 
 class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
-  var onTapRecognizer;
-
   TextEditingController textEditingController = TextEditingController();
   // ..text = "123456";
 
-  StreamController<ErrorAnimationType> errorController;
+  // ignore: close_sinks
+  StreamController<ErrorAnimationType>? errorController;
 
   bool hasError = false;
   String currentText = "";
-  final GlobalKey<ScaffoldState> scaffoldKey = GlobalKey<ScaffoldState>();
   final formKey = GlobalKey<FormState>();
 
   @override
   void initState() {
-    onTapRecognizer = TapGestureRecognizer()
-      ..onTap = () {
-        Navigator.pop(context);
-      };
     errorController = StreamController<ErrorAnimationType>();
     super.initState();
   }
 
   @override
   void dispose() {
-    errorController.close();
+    errorController!.close();
 
     super.dispose();
+  }
+
+  // snackBar Widget
+  snackBar(String? message) {
+    return ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message!),
+        duration: const Duration(seconds: 2),
+      ),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.blue.shade50,
-      key: scaffoldKey,
+      backgroundColor: Constants.primaryColor,
       body: GestureDetector(
         onTap: () {},
-        child: Container(
+        child: SizedBox(
           height: MediaQuery.of(context).size.height,
           width: MediaQuery.of(context).size.width,
           child: ListView(
             children: <Widget>[
-              SizedBox(height: 30),
-              Container(
+              const SizedBox(height: 30),
+              SizedBox(
                 height: MediaQuery.of(context).size.height / 3,
-                child: FlareActor(
-                  "assets/otp.flr",
-                  animation: "otp",
-                  fit: BoxFit.fitHeight,
-                  alignment: Alignment.center,
+                child: ClipRRect(
+                  borderRadius: BorderRadius.circular(30),
+                  child: Image.asset(Constants.otpGifImage),
                 ),
               ),
-              SizedBox(height: 8),
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8.0),
+              const SizedBox(height: 8),
+              const Padding(
+                padding: EdgeInsets.symmetric(vertical: 8.0),
                 child: Text(
                   'Phone Number Verification',
                   style: TextStyle(fontWeight: FontWeight.bold, fontSize: 22),
@@ -101,17 +110,18 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                       text: "Enter the code sent to ",
                       children: [
                         TextSpan(
-                            text: widget.phoneNumber,
-                            style: TextStyle(
+                            text: "${widget.phoneNumber}",
+                            style: const TextStyle(
                                 color: Colors.black,
                                 fontWeight: FontWeight.bold,
                                 fontSize: 15)),
                       ],
-                      style: TextStyle(color: Colors.black54, fontSize: 15)),
+                      style:
+                          const TextStyle(color: Colors.black54, fontSize: 15)),
                   textAlign: TextAlign.center,
                 ),
               ),
-              SizedBox(
+              const SizedBox(
                 height: 20,
               ),
               Form(
@@ -126,11 +136,15 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                         fontWeight: FontWeight.bold,
                       ),
                       length: 6,
-                      obscureText: false,
+                      obscureText: true,
                       obscuringCharacter: '*',
+                      obscuringWidget: const FlutterLogo(
+                        size: 24,
+                      ),
+                      blinkWhenObscuring: true,
                       animationType: AnimationType.fade,
                       validator: (v) {
-                        if (v.length < 3) {
+                        if (v!.length < 3) {
                           return "I'm from validator";
                         } else {
                           return null;
@@ -139,20 +153,17 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                       pinTheme: PinTheme(
                         shape: PinCodeFieldShape.box,
                         borderRadius: BorderRadius.circular(5),
-                        fieldHeight: 60,
-                        fieldWidth: 50,
-                        activeFillColor:
-                            hasError ? Colors.orange : Colors.white,
+                        fieldHeight: 50,
+                        fieldWidth: 40,
+                        activeFillColor: Colors.white,
                       ),
                       cursorColor: Colors.black,
-                      animationDuration: Duration(milliseconds: 300),
-                      textStyle: TextStyle(fontSize: 20, height: 1.6),
-                      backgroundColor: Colors.blue.shade50,
+                      animationDuration: const Duration(milliseconds: 300),
                       enableActiveFill: true,
                       errorAnimationController: errorController,
                       controller: textEditingController,
                       keyboardType: TextInputType.number,
-                      boxShadows: [
+                      boxShadows: const [
                         BoxShadow(
                           offset: Offset(0, 1),
                           color: Colors.black12,
@@ -160,19 +171,19 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                         )
                       ],
                       onCompleted: (v) {
-                        print("Completed");
+                        debugPrint("Completed");
                       },
                       // onTap: () {
                       //   print("Pressed");
                       // },
                       onChanged: (value) {
-                        print(value);
+                        debugPrint(value);
                         setState(() {
                           currentText = value;
                         });
                       },
                       beforeTextPaste: (text) {
-                        print("Allowing to paste $text");
+                        debugPrint("Allowing to paste $text");
                         //if you return true then it will show the paste confirmation dialog. Otherwise if false, then nothing will happen.
                         //but you can show anything you want here, like your pop up saying wrong paste format or etc
                         return true;
@@ -183,31 +194,35 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                 padding: const EdgeInsets.symmetric(horizontal: 30.0),
                 child: Text(
                   hasError ? "*Please fill up all the cells properly" : "",
-                  style: TextStyle(
+                  style: const TextStyle(
                       color: Colors.red,
                       fontSize: 12,
                       fontWeight: FontWeight.w400),
                 ),
               ),
-              SizedBox(
+              const SizedBox(
                 height: 20,
               ),
-              RichText(
-                textAlign: TextAlign.center,
-                text: TextSpan(
-                    text: "Didn't receive the code? ",
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  const Text(
+                    "Didn't receive the code? ",
                     style: TextStyle(color: Colors.black54, fontSize: 15),
-                    children: [
-                      TextSpan(
-                          text: " RESEND",
-                          recognizer: onTapRecognizer,
-                          style: TextStyle(
-                              color: Color(0xFF91D3B3),
-                              fontWeight: FontWeight.bold,
-                              fontSize: 16))
-                    ]),
+                  ),
+                  TextButton(
+                    onPressed: () => snackBar("OTP resend!!"),
+                    child: const Text(
+                      "RESEND",
+                      style: TextStyle(
+                          color: Color(0xFF91D3B3),
+                          fontWeight: FontWeight.bold,
+                          fontSize: 16),
+                    ),
+                  )
+                ],
               ),
-              SizedBox(
+              const SizedBox(
                 height: 14,
               ),
               Container(
@@ -215,30 +230,27 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                     const EdgeInsets.symmetric(vertical: 16.0, horizontal: 30),
                 child: ButtonTheme(
                   height: 50,
-                  child: FlatButton(
+                  child: TextButton(
                     onPressed: () {
-                      formKey.currentState.validate();
+                      formKey.currentState!.validate();
                       // conditions for validating
-                      if (currentText.length != 6 || currentText != "towtow") {
-                        errorController.add(ErrorAnimationType
+                      if (currentText.length != 6 || currentText != "123456") {
+                        errorController!.add(ErrorAnimationType
                             .shake); // Triggering error shake animation
-                        setState(() {
-                          hasError = true;
-                        });
+                        setState(() => hasError = true);
                       } else {
-                        setState(() {
-                          hasError = false;
-                          scaffoldKey.currentState.showSnackBar(SnackBar(
-                            content: Text("Aye!!"),
-                            duration: Duration(seconds: 2),
-                          ));
-                        });
+                        setState(
+                          () {
+                            hasError = false;
+                            snackBar("OTP Verified!!");
+                          },
+                        );
                       }
                     },
                     child: Center(
                         child: Text(
                       "VERIFY".toUpperCase(),
-                      style: TextStyle(
+                      style: const TextStyle(
                           color: Colors.white,
                           fontSize: 18,
                           fontWeight: FontWeight.bold),
@@ -251,32 +263,36 @@ class _PinCodeVerificationScreenState extends State<PinCodeVerificationScreen> {
                     boxShadow: [
                       BoxShadow(
                           color: Colors.green.shade200,
-                          offset: Offset(1, -2),
+                          offset: const Offset(1, -2),
                           blurRadius: 5),
                       BoxShadow(
                           color: Colors.green.shade200,
-                          offset: Offset(-1, 2),
+                          offset: const Offset(-1, 2),
                           blurRadius: 5)
                     ]),
               ),
-              SizedBox(
+              const SizedBox(
                 height: 16,
               ),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  FlatButton(
-                    child: Text("Clear"),
+                  Flexible(
+                      child: TextButton(
+                    child: const Text("Clear"),
                     onPressed: () {
                       textEditingController.clear();
                     },
-                  ),
-                  FlatButton(
-                    child: Text("Set Text"),
+                  )),
+                  Flexible(
+                      child: TextButton(
+                    child: const Text("Set Text"),
                     onPressed: () {
-                      textEditingController.text = "123456";
+                      setState(() {
+                        textEditingController.text = "123456";
+                      });
                     },
-                  ),
+                  )),
                 ],
               )
             ],