+++ test/pin_code_fields_test.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -7,7 +6,8 @@ import 'package:pin_code_fields/pin_code_fields.dart';
 
 void main() {
   final TestWidgetsFlutterBinding binding =
-      TestWidgetsFlutterBinding.ensureInitialized();
+      TestWidgetsFlutterBinding.ensureInitialized()
+          as TestWidgetsFlutterBinding;
 
   testWidgets('disposes error stream', (WidgetTester tester) async {
     final StreamController<ErrorAnimationType> controller =